Need: 1.subclass. 2.form use label name in annotation, if empty, use field name. Done: contextmenu for root-node copy&delete. Doing: need to check if the variable is list or not. Done: store real number in sessionStorage, and can output.

+++ src/main/java/jetty/demo/PersonDemo.java
@@ -18,7 +18,7 @@ public class PersonDemo implements Serializable{
 	@AnnotationForm(
 			style = { @AnnotationStyle(input=InputTypeControl.text, value = "1") },
 			name = "age")
-	private int age;
+	private long age;
 	
 	@AnnotationForm(
 			style = { @AnnotationStyle(input=InputTypeControl.text, value = "huang") },
@@ -60,7 +60,7 @@ public class PersonDemo implements Serializable{
 		this.age = age;
 	}
 	
-	public int getAge() {
+	public long getAge() {
 		return this.age;
 	}
